Extensive refactoring for common code
- Move TAP_TOGGLE logic into user/tap_tog.c. State shared via extern
  vars.
- Consolidate common config and rules into user folder
- Clean up keymaps and remove unused macros and functions
- Enable link time optimization to free up space in builds
