+++ ergodox_ez/rules.mk
@@ -3,8 +3,5 @@ ifndef QUANTUM_DIR
     include ../../../../Makefile
 endif
 
-# disable magic commands
-COMMAND_ENABLE = no
-
 # enable hid_listen support (400kb)
 CONSOLE_ENABLE = yes
