+++ user/config.h
@@ -1,9 +1,17 @@
 #pragma once
 
+// how long before a tap becomes a hold
+#undef TAPPING_TERM
+#define TAPPING_TERM 100
+
+// makes tap and hold keys work better for fast typers who don't want
+// tapping term set above 500
 #define PERMISSIVE_HOLD
 
-#undef OVERWATCH_THEME
-#define OVERWATCH_THEME                        \
-    WD_NOTE(_A6),  Q__NOTE(_E6), Q__NOTE(_A6), \
-    WD_NOTE(_B6),  Q__NOTE(_E6), Q__NOTE(_B6), \
-    B__NOTE(_CS7),
+// tap anyway, even after TAPPING_TERM, if there was no other key
+// interruption between press and release
+#define RETRO_TAPPING
+
+// how many taps before triggering the toggle
+#undef ONESHOT_TAP_TOGGLE
+#define ONESHOT_TAP_TOGGLE 2
